scraper_function/requirements.txt
azure-functions==1.2.1
pymongo==3.10.1
beautifulsoup4==4.9.1
bs4==0.0.1
urllib3==1.25.9
certifi==2020.4.5.2